+++ src/mintSDL/display/texture.h
@@ -21,7 +21,7 @@ MintTexture* mint_TextureFromPNG(SDL_Renderer* renderer, char* path);
 MintTexture* mint_TextureFromText(SDL_Renderer* renderer, TTF_Font* font, char* text, SDL_Color colour);
 MintTexture* mint_TextureFromSurface(SDL_Renderer* renderer, SDL_Surface* surface);
 
-void mint_TextureUpdate(MintTexture* mintTexture, double elapsed);
+void mint_TextureUpdate(MintTexture* mintTexture, float elapsed);
 void mint_TextureRender(MintTexture* mintTexture);
 void mint_TextureFree(MintTexture* mintTexture);
