+++ src/mintSDL/maths/phys.cpp
@@ -41,9 +41,9 @@ MintPhys* mint_PhysCreate(MintTexture* mintTexture, MintPhysWorld* physWorld, ch
 }
 
 void mint_PhysSetGravity(MintPhysWorld* world, float gravityX, float gravityY) { world->world->SetGravity({ gravityX, gravityY }); }
-void mint_PhysStepWorld(MintPhysWorld* world, double elapsed) { world->world->Step((float)elapsed, 6, 2); }
-float mint_PhysMetreToPixel(double metre) { return (float)(metre * 100); }
-float mint_PhysPixelToMetre(double pixel) { return (float)(pixel / 100); }
+void mint_PhysStepWorld(MintPhysWorld* world, float elapsed) { world->world->Step((float)elapsed, 6, 2); }
+float mint_PhysMetreToPixel(float metre) { return (float)(metre * 100); }
+float mint_PhysPixelToMetre(float pixel) { return (float)(pixel / 100); }
 
 void mint_PhysApplyForce(MintPhys* phys, float forceX, float forceY)
 {
@@ -55,7 +55,7 @@ void mint_PhysSetVelocity(MintPhys* phys, float veloX, float veloY)
 	phys->body->SetLinearVelocity(b2Vec2(veloX, veloY));
 }
 
-void mint_PhysUpdate(MintPhys* phys, double elapsed)
+void mint_PhysUpdate(MintPhys* phys, float elapsed)
 {
 	b2Vec2 pos = phys->body->GetPosition();
 	pos.x = mint_PhysMetreToPixel(pos.x);
